@@ -34,38 +34,37 @@
         "separation logic",
         "automatic deduction"
     ],
-    "dependencies": {
+    "dependencies": {},
+    "devDependencies": {
+        "@types/glob-to-regexp": "^0.4.1",
+        "@types/mocha": "^9.0.0",
+        "@types/node": "^17.0.9",
+        "@types/rimraf": "^3.0.2",
+        "@types/tmp": "^0.2.3",
+        "@types/vscode": "^1.63.1",
+        "@typescript-eslint/eslint-plugin": "^5.10.0",
+        "@typescript-eslint/parser": "^5.10.0",
+        "eslint": "^8.7.0",
+        "eslint-webpack-plugin": "^3.1.1",
+        "mocha": "^9.1.4",
+        "npm-run-all": "^4.1.5",
+        "tmp": "^0.2.1",
+        "ts-loader": "^9.2.6",
+        "typescript": "^4.5.4",
+        "vsce": "^2.6.3",
+        "vscode-test": "^1.6.1",
+        "webpack": "^5.66.0",
+        "webpack-cli": "^4.9.1",
         "glob-to-regexp": "^0.4.1",
-        "node-ipc": "^9.1.4",
+        "node-ipc": "^10.1.0",
         "rimraf": "^3.0.2",
-        "strip-json-comments": "3.1.1",
+        "strip-json-comments": "4.0.0",
         "unused-filename": "^3.0.0",
         "viz.js": "^2.1.2",
         "vscode-debugadapter": "^1.34.0",
         "vscode-debugprotocol": "^1.34.0",
         "vscode-languageclient": "^5.2.1",
-        "vscode-uri": "^3.0.2"
-    },
-    "devDependencies": {
-        "@types/glob-to-regexp": "^0.4.1",
-        "@types/mocha": "^8.2.3",
-        "@types/node": "^15.6.1",
-        "@types/rimraf": "^3.0.0",
-        "@types/tmp": "^0.2.0",
-        "@types/vscode": "^1.43.0",
-        "@typescript-eslint/eslint-plugin": "^4.28.3",
-        "@typescript-eslint/parser": "^4.26.1",
-        "eslint": "^7.28.0",
-        "eslint-webpack-plugin": "^2.5.4",
-        "mocha": "^8.1.3",
-        "npm-run-all": "^4.1.5",
-        "tmp": "^0.2.1",
-        "ts-loader": "^9.2.2",
-        "typescript": "^4.2.4",
-        "vsce": "^1.88.0",
-        "vscode-test": "^1.3.0",
-        "webpack": "^5.44.0",
-        "webpack-cli": "^4.0.0"
+        "vscode-uri": "^3.0.3"
     },
     "scripts": {
         "webpack-production": "webpack --mode production",
@@ -73,11 +72,12 @@
         "webpack-development-watch": "webpack --mode development --watch",
         "tsc": "tsc",
         "clean": "rimraf dist",
-        "compileServer": "cd ../server && npm run compile && cd ../client",
-        "compile": "npm-run-all --sequential clean compileServer webpack-development",
-        "watch": "npm-run-all --sequential clean compileServer webpack-development-watch",
-        "vscode:prepublish": "npm-run-all --sequential clean compileServer webpack-production",
-        "test-compile": "npm-run-all --sequential clean compileServer tsc",
+        "compile-server": "cd ../server && npm run compile && cd ../client",
+        "compile-server-production": "cd ../server && npm run compile-production && cd ../client",
+        "compile": "npm-run-all --sequential clean compile-server webpack-development",
+        "watch": "npm-run-all --sequential clean compile-server webpack-development-watch",
+        "vscode:prepublish": "npm-run-all --sequential clean compile-server-production webpack-production",
+        "test-compile": "npm-run-all --sequential clean compile-server tsc",
         "pretest": "npm run test-compile",
         "test": "node ./dist/test/runTest.js",
         "package": "vsce package"
