@@ -12,37 +12,41 @@
         "vscode": "^1.45.0",
         "node": "*"
     },
-    "dependencies": {
-        "@octokit/rest": "^18.5.3",
+    "devDependencies": {
+        "@octokit/rest": "^18.12.0",
+        "@types/mocha": "^9.0.0",
+        "@types/node": "^17.0.9",
+        "@types/node-ipc": "^9.2.0",
+        "@types/stream-json": "^1.7.2",
+        "@types/vscode": "^1.63.1",
+        "@types/yargs": "^17.0.8",
+        "@vercel/ncc": "^0.33.1",
         "extract-zip": "^2.0.1",
         "fs-extra": "^10.0.0",
-        "glob-to-regexp": "^0.4.0",
-        "got": "^11.8.2",
+        "glob-to-regexp": "^0.4.1",
+        "got": "^11.7.0",
         "locate-java-home": "git+https://github.com/viperproject/locate-java-home.git",
-        "node-ipc": "^10.0.2",
-        "request": "^2.88.2",
-        "sound-play": "^1.1.0",
-        "stream-json": "^1.7.1",
-        "tree-kill": "^1.2.2",
-        "vscode-languageserver": "^5.2.1",
-        "vscode-uri": "^3.0.2",
-        "yargs": "^17.0.1"
-    },
-    "devDependencies": {
-        "@types/mocha": "^8.0.3",
-        "@types/node": "^15.6.1",
-        "@types/node-ipc": "^9.1.4",
-        "@types/stream-json": "^1.5.1",
-        "@types/vscode": "^1.58.0",
-        "@types/yargs": "^17.0.2",
-        "@vercel/ncc": "^0.28.6",
+        "node-ipc": "^10.1.0",
         "npm-run-all": "^4.1.5",
+        "request": "^2.88.2",
         "rimraf": "^3.0.2",
-        "typescript": "^4.2.4"
+        "sound-play": "^1.1.0",
+        "stream-json": "^1.7.3",
+        "tree-kill": "^1.2.2",
+        "ts-loader": "^9.2.6",
+        "typescript": "^4.5.4",
+        "vscode-languageserver": "^5.2.1",
+        "vscode-uri": "^3.0.3",
+        "webpack": "^5.66.0",
+        "webpack-cli": "^4.9.1",
+        "yargs": "^17.3.1"
     },
     "scripts": {
         "clean": "rimraf dist",
-        "bundle": "ncc build src/server.ts --source-map --license licenses.txt --out ../client/server",
-        "compile": "npm-run-all --sequential clean bundle"
+        "webpack-production": "webpack --mode production",
+        "webpack-development": "webpack --mode development",
+        "webpack-development-watch": "webpack --mode development --watch",
+        "compile": "npm-run-all --sequential clean webpack-development",
+        "compile-production": "npm-run-all --sequential clean webpack-production"
     }
 }
